Add files via upload
Pegasus - The one creature from Greek mythology that I could find that would try to protect humans. Pegasus, that is his name, he is the child of Medusa and Poseidon from the myth of Medusa. If you want to refer to the species of winged horses those are pegasi. Chiron is the other one that would be friendly and he is a centaur, despite what Disney wants you to believe.